build refinement, refactoring
- added webpack loader for lezer grammar
- moved highlighting to a separate module
- fixed broken diagnostic state
- memo-ed language support
