+++ packages/components/package.json
@@ -56,6 +56,7 @@
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.3.1",
     "jsdom": "^20.0.2",
+    "lezer-loader": "^0.3.0",
     "mini-css-extract-plugin": "^2.6.1",
     "postcss-cli": "^10.0.0",
     "postcss-import": "^15.0.0",
