+++ packages/components/src/components/grammar/squiggle.grammar.d.ts
@@ -1,0 +1,4 @@
+import { LRParser } from "@lezer/lr";
+
+declare const parser: LRParser;
+export default parser;
