--- packages/components/src/components/grammar/squiggle.terms.ts
@@ -1,12 +1,0 @@
-// This file was generated by lezer-generator. You probably shouldn't edit it.
-export const
-  spaces = 73,
-  Program = 1,
-  Binding = 2,
-  String = 5,
-  Boolean = 6,
-  Number = 8,
-  LambdaArgs = 11,
-  AccessIdentifier = 24,
-  newline = 83,
-  Statement = 64
