+++ packages/components/src/components/grammar/squiggle.ts
@@ -1,18 +1,64 @@
-// This file was generated by lezer-generator. You probably shouldn't edit it.
-import {LRParser} from "@lezer/lr"
-const spec_identifier = {__proto__:null,true:14, false:14, if:34, then:36, else:38}
-export const parser = LRParser.deserialize({
-  version: 14,
-  states: "-tQYQPOOO}QPOOO!oQPO'#C_OOQO'#Cb'#CbO#yQQO'#CeO}QPO'#CpO$OQPO'#CtOOQO'#Cs'#CsQ&gQPOOO&nQPO'#C{OOQO'#C|'#C|OOQO'#DO'#DOO}QPO'#C}OOQO'#Dz'#DzO&uQPO'#C^O&zQPO'#DoOOQO'#Do'#DoQYQPOOO}QPO'#ClOOQO'#Cu'#CuQ%iQPOOO'PQQO,59PO'XQPO,59[O'`QPO'#DqO'eQPO,59`O)OQPO,59dOOQO'#DR'#DROOQO'#DT'#DTOOQO'#DV'#DVOOQO'#DX'#DXOOQO'#DZ'#DZOOQO'#D]'#D]OOQO'#D_'#D_O}QPO,59lO}QPO,59lO}QPO,59lOOQO'#Db'#DbO}QPO,59{OOQO'#De'#DeOOQO'#Dg'#DgOOQO'#Di'#DiOOQO'#Dk'#DkOOQO'#Dm'#DmO}QPO,5:OO}QPO,5:OO}QPO,5:OOOQO'#Dt'#DtQ)VQPOOO)[QPO'#EOO)fQPO,59gO)kQPO,59iO}QPO,58xOOQO,5:Z,5:ZOOQO-E7m-E7mO+RQPO,59WOOQO'#Ch'#ChO+YQQO'#CgO+bQQO1G.kOOQO1G.v1G.vOOQO'#Cw'#CwOOQO,5:],5:]OOQO-E7o-E7oO+gQPO'#CyO+qQPO'#D|O+yQPO1G/OO-cQPO1G/WO-jQPO1G/WO.YQPO1G/WO/]QPO1G/gO2WQPO1G/jO2_QPO1G/jO2wQPO1G/jOOQO-E7r-E7rO}QPO'#DsO4QQPO,5:jOOQO1G/R1G/RO4YQPO1G.dPOQO'#C_'#C_O}QPO1G.rO4aQPO'#DpO4fQQO,59RO}QPO7+$VO}QPO'#DrO4nQPO,5:hOOQO7+$j7+$jO4vQPO,5:_OOQO-E7q-E7qO5QQPO7+$^OOQO'#Cj'#CjOOQO,5:[,5:[OOQO-E7n-E7nO5XQPO<<GqO5`QPO'#CzOOQO,5:^,5:^OOQO-E7p-E7pO}QPO<<GxOOQOAN=]AN=]O5jQPOAN=d",
-  stateData: "6m~O!kOS!lOS~OT]OVROW]OYSOabOeTOsZO!mQO!qXO!tYO!u`O~OT]OVROW]OYSOabOeTOsZO!mcO!qXO!tYO~OSRXeiXjiXsiXviXxiXziX|iX!OiX!QiX!SiX!ViX!YiX![iX!^iX!`iX!iiX!uiX~O!oeO~OjgOehXshXvhXxhXzhX|hX!OhX!QhX!ShX!VhX!YhX![hX!^hX!`hX!ihX!uhXfhX]hX!shXbhXchX_hX~OeiOszOvjOxkOzlO|mO!OnO!QoO!SpO!VtO!YvO![wO!^xO!`yO~O!u!OO~P%iO!s!rP~P}OS!TO~O!u!UO~O!m!XO!oZP~Of![O~P%iO!m!]O~OjgOehashavhaxhazha|ha!Oha!Qha!Sha!Vha!Yha![ha!^ha!`ha!iha!uhafha]ha!shabhacha_ha~Of!pP~P}O!u!OO~O]!kO!s!rX~P%iO!s!mO~OeiOsqavqaxqazqa|qa!Oqa!Qqa!Sqa!Vqa!Yqa![qa!^qa!`qa!iqa!uqafqa]qa!sqabqacqa_qa~Ob!pO~P%iO]!qO!oZX~O!o!sO~O]mXfmX~P%iO]!tOf!pX~Of!vO~OeiOszOzlO|mO!OnO!QoO!SpO!VtO!YvO![wO!^xO!`yOvti!iti!utifti]ti!stibticti_ti~OxkO~P,OOxti~P,OOeiOszO!VtO!YvO![wO!^xO!`yO~Ovtixtizti|ti!Oti!Qti!Sti!iti!utifti]ti!stibticti_ti~P-qOeiO!^xOs!Tiv!Tix!Tiz!Ti|!Ti!O!Ti!Q!Ti!S!Ti!V!Ti!Y!Ti![!Ti!`!Ti!i!Ti!u!Tif!Ti]!Ti!s!Tib!Tic!Ti_!Ti~OeiOs!Wiv!Wix!Wiz!Wi|!Wi!O!Wi!Q!Wi!S!Wi!V!Wi!Y!Wi![!Wi!`!Wi!i!Wi!u!Wif!Wi]!Wi!s!Wib!Wic!Wi_!Wi~O!^xO~P0sO!^!Wi~P0sOeiO!VtO!YvO![wO!^xO~Os!Wiv!Wix!Wiz!Wi|!Wi!O!Wi!Q!Wi!S!Wi!`!Wi!i!Wi!u!Wif!Wi]!Wi!s!Wib!Wic!Wi_!Wi~P2fO]!kO!s!ra~O!uQi~P%iO!m!zO~O]!qO!oZa~O]!tOf!pa~O]!ga!s!ga~P%iOc#RO~P%iO_#SO~P%iO]nXfnX~P%iOv`!Rx`!Rz`!R|`!R!O`!R!Q`!R!S`!R!i`!R!u`!Rf`!R]`!R!s`!Rb`!Rc`!R_`!R~P-qO",
-  goto: "*S!sPP!t!xPP!|PP!|P#e#hP#kP!|PPP!|PP!|#n$VP$n!|$q$t!|!|!|$wP!|%`P%mP%{P%{P%{P%{P%{P!|&[P!|&nP&nP'QP'fP'fP'w'}(T(Z(a(gPPPPP(mP)|P*PT_OaT^Oa{]OPTX[abiqrsu{|}!T!k!p!s!t#RR!ZeR!YeR!{!q{VOPTX[abiqrsu{|}!T!k!p!s!t#R{UOPTX[abiqrsu{|}!T!k!p!s!t#RR!^gR!aiR#P!t{[OPTX[abiqrsu{|}!T!k!p!s!t#RgqWdf!Q!W!`!n!w!y!}#OirWdf!Q!W!`!c!n!w!y!}#OksWdf!Q!W!`!c!d!n!w!y!}#OquWdf!Q!W!`!c!d!e!i!n!w!y!}#O#Tq{Wdf!Q!W!`!c!d!e!i!n!w!y!}#O#Tu|Wdf!Q!W!`!c!d!e!f!g!i!n!w!y!}#O#To}Wdf!Q!W!`!c!d!e!n!w!y!}#O#TQaOR!VaQ!r!YR!|!rQhUR!_hQ!u!aR#Q!uQ!l!QR!x!lQ!PWR!j!PSWOaQdPQfTQ!QXQ!S[Q!WbQ!`iQ!cqQ!drQ!esQ!fuQ!g{Q!h|Q!i}Q!n!TQ!w!kQ!y!pQ!}!sQ#O!tR#T#RR!biR!RX",
-  nodeNames: "⚠ Program Binding VariableName Equals String Boolean Boolean Number LambdaExpr { LambdaArgs LambdaParameter , LambdaParameter } IfExpr if then else ParenExpr ( ) AccessExpr AccessIdentifier Field . Field CallExpr Argument Argument ArrayExpr Void UnaryExpr ArithOp - LogicExpr LogicOp || LogicOp && LogicOp > LogicOp < LogicOp <= LogicOp >= LogicOp == ControlExpr ControlOp |> ArithExpr ArithOp * ArithOp / ArithOp ^ ArithOp + ArithOp Statement",
-  maxTerm: 83,
-  skippedNodes: [0],
-  repeatNodeCount: 6,
-  tokenData: "*b~RoYZ#Spq#Xrs#ast$Otu$gvw%[wx%gxy&Pyz&^z{&c{|&h|}&m}!O&r!O!P&w!P!Q(R!Q![(W!^!_(i!_!`(v!`!a)T!c!}$g!}#O)b#P#Q)g#Q#R)l#R#S$g#T#o$g#o#p)q#p#q)v#q#r*]$g;'S$g;'S;=`%U<%lO$g~#XO!u~~#^P!k~pq#X~#dTOr#ars#ss;'S#a;'S;=`#x<%lO#a~#xOT~~#{P;=`<%l#a~$TS!l~OY$OZ;'S$O;'S;=`$a<%lO$O~$dP;=`<%l$O~$lW!m~tu$g!Q![$g!c!}$g#R#S$g#T#o$g$g;'S$g;'S;=`%U<%lO$g~%XP;=`<%l$g~%_Pvw%b~%gOx~~%jTOw%gwx#sx;'S%g;'S;=`%y<%lO%g~%|P;=`<%l%g~&UPe~yz&X~&^O!t~~&cOf~~&hO!Y~~&mO!`~~&rO]~~&wOs~~'ORj~W~!Q!['X!g!h'g#X#Y'g~'^RW~!Q!['X!g!h'g#X#Y'g~'jR{|'s}!O's!Q!['y~'vP!Q!['y~(OPW~!Q!['y~(WO![~~(]SW~!O!P'X!Q![(W!g!h'g#X#Y'g~(nP|~!_!`(q~(vO!O~~({PS~!_!`)O~)TO!S~~)YPz~!_!`)]~)bO!Q~~)gO!q~~)lO!s~~)qO!^~~)vOY~R){Q!oQ!`!a*R#p#q*WP*WO!VPP*]OvP~*bO_~",
-  tokenizers: [0, 1],
-  topRules: {"Program":[0,1],"Statement":[1,64]},
-  specialized: [{term: 75, get: value => spec_identifier[value] || -1}],
-  tokenPrec: 0
-})
+import { LRLanguage, LanguageSupport } from "@codemirror/language";
+import { styleTags, tags as t } from "@lezer/highlight";
+import { snippetCompletion } from "@codemirror/autocomplete";
+import parser from "./squiggle.grammar";
+
+const squiggleLang = LRLanguage.define({
+  name: "squiggle",
+  parser: parser.configure({
+    props: [
+      styleTags({
+        if: t.keyword,
+        then: t.keyword,
+        else: t.keyword,
+
+        Equals: t.definitionOperator,
+
+        ArithOp: t.arithmeticOperator,
+        LogicOp: t.logicOperator,
+        ControlOp: t.controlOperator,
+
+        "{ }": t.brace,
+        "[ ]": t.squareBracket,
+        "( )": t.paren,
+        "|": t.squareBracket,
+        ",": t.separator,
+
+        Syntax: t.annotation,
+        Boolean: t.bool,
+        Number: t.integer,
+        String: t.string,
+        Comment: t.comment,
+        Void: t.escape,
+
+        FunctionName: t.function(t.propertyName),
+
+        LambdaSyntax: t.blockComment,
+
+        VariableName: t.variableName,
+        Field: t.variableName,
+        LambdaParameter: t.variableName,
+      }),
+    ],
+  }),
+  languageData: {
+    commentTokens: {
+      line: "#",
+    },
+    autocomplete: [
+      snippetCompletion("{ |${args}| ${body} ", {
+        label: "{",
+        detail: "lambda function",
+      }),
+    ],
+    closeBrackets: {
+      brackets: ['"', "'", "(", "{", "|"],
+    },
+  },
+});
+
+function squiggle() {
+  return new LanguageSupport(squiggleLang, []);
+}
+
+export default squiggle
